requirements.txt
# Pinned runtime dependencies for the app
Flask==3.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
bleach==6.1.0
gunicorn==21.2.0
Pillow==10.1.0
reportlab==4.0.7
weasyprint==61.2
# google-genai client used across utils/*
google-genai==1.0.0
